requirements.txt
torch>=2.0.0                    # PyTorch for TTS model
torchaudio>=2.0.0              # Audio processing
spacy>=3.7.0,<3.8.0           # NLP and sentence segmentation
openai-whisper>=20231117       # Speech-to-text validation
auto-editor>=22.39.1,<23.0.0   # Audio post-processing (compatible with numpy 1.26)
chatterbox-tts>=0.1.2          # TTS Generation - Updated to latest version with min_p parameters
fuzzywuzzy>=0.18.0             # Fuzzy text matching
python-Levenshtein>=0.21.0    # String similarity calculations
pydantic>=2.0.0                # Data validation
pyyaml>=6.0                    # Configuration management
tqdm>=4.65.0                   # Progress tracking
pytest>=7.0.0                  # Testing framework
pytest-cov>=4.0.0             # Coverage reporting